Refactor: move arrayToAnalytics() to tsutls
This is the last dependency of aggregator on data-broker
