+++ packages/aggregator/src/__tests__/analytics/aggregateAnalytics/aggregations/offsetPeriod.test.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { arrayToAnalytics } from '@tupaia/data-broker';
+import { arrayToAnalytics } from '@tupaia/tsutils';
 import {
   getDateRangeForOffsetPeriod,
   offsetPeriod,
