+++ packages/data-broker/src/utils.ts
@@ -3,10 +3,6 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { Analytic } from './types';
-
-type ArrayAnalytic = [string, string, string, string | number];
-
 const DATA_ELEMENT = 'dataElement';
 const DATA_GROUP = 'dataGroup';
 const SYNC_GROUP = 'syncGroup';
@@ -22,11 +18,3 @@ export const EMPTY_ANALYTICS_RESULTS = {
     dataElementCodeToName: {},
   },
 };
-
-export const arrayToAnalytics = (arrayAnalytics: ArrayAnalytic[]): Analytic[] =>
-  arrayAnalytics.map(([dataElement, organisationUnit, period, value]) => ({
-    dataElement,
-    organisationUnit,
-    period,
-    value,
-  }));
