+++ packages/tsutils/src/__tests__/arrayToAnalytics.test.ts
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { arrayToAnalytics } from '../utils';
+import { arrayToAnalytics } from '..';
 
 describe('utils', () => {
   describe('arrayToAnalytic', () => {
