+++ packages/tsutils/src/arrayToAnalytics.ts
@@ -1,0 +1,21 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
+
+interface Analytic {
+  dataElement: string;
+  organisationUnit: string;
+  period: string; // should be in format YYYYMMDD e.g. "20210103"
+  value: string | number;
+}
+
+type ArrayAnalytic = [string, string, string, string | number];
+
+export const arrayToAnalytics = (arrayAnalytics: ArrayAnalytic[]): Analytic[] =>
+  arrayAnalytics.map(([dataElement, organisationUnit, period, value]) => ({
+    dataElement,
+    organisationUnit,
+    period,
+    value,
+  }));
