+++ packages/web-config-server/src/__tests__/apiV1/dataBuilders/helpers/calculateOperationForAnalytics.test.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
  */
 
-import { arrayToAnalytics } from '@tupaia/data-broker';
+import { arrayToAnalytics } from '@tupaia/tsutils';
 import { NO_DATA_AVAILABLE } from '/apiV1/dataBuilders/constants';
 
 import { calculateOperationForAnalytics } from '/apiV1/dataBuilders/helpers';
