+++ packages/web-config-server/src/__tests__/apiV1/utils/layerYearOnYear.test.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { arrayToAnalytics } from '@tupaia/data-broker';
+import { arrayToAnalytics } from '@tupaia/tsutils';
 import { layerYearOnYear } from '../../../apiV1/utils/layerYearOnYear';
 
 describe('layerYearOnYear()', () => {
